moved spark snowflake to diff connector
